AI BASED RESUME SCREENING/code/requirements.txt
supabase
pdfplumber
pdfminer.six
PyMuPDF
python-docx
sentence-transformers
faiss-cpu
numpy
torch
rapidfuzz
pandas
groq
chardet<6
requests>=2.32.3
pytesseract
pillow
pdf2image
flask
flask-cors
